modify code by regulation
