+++ Hyper_tuner/README.md
@@ -1,4 +1,4 @@
-# Tuning_kit demo
+# Hyper Tuner demo
 鲲鹏Hyper Tuner系统性能优化工具是针对TaiShan服务器的性能分析和优化工具，能收集服务器的处理器硬件、操作系统、进程/线程、函数等各个层次的性能数据，分析出系统性能指标，定位到系统瓶颈及热点函数，并给出相应解决问题的优化建议。
 这个demo展示了如何使用鲲鹏Hyper Tuner系统性能优化工具中的各个功能模块儿发现系统或应用中存在的问题，解决系统瓶颈并对代码进行优化。
 \
