+++ Hyper_tuner/testdemo/cache/cahce_hit/cache_hit.c
@@ -31,7 +31,7 @@ gcc -g cache_hit.c -o cache_hit
 int main(void)
 {
     int i, j;
-    int *p = (int*) malloc(sizeof(int) * N)
+    int *p = (int*) malloc(sizeof(int) * N);
     double msecs;
     struct timeval before, after;
