+++ Hyper_tuner/testdemo/lock/pthread_mutex_long_mod.c
@@ -14,23 +14,20 @@
  * limitations under the License.
  */
 
-// gcc -g pthread_mutex.c -o pthread_mutex -lpthread
+// gcc -g pthread_mutex_long_mod.c -o pthread_mutex_long_mod -lpthread -lm
 #include <pthread.h>
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
 #include <errno.h>
 #include <sys/timeb.h>
+#include <math.h>
+
+#define N    100000
 
 static long num = 1;
-static long count = 10000000000;
+static long count = 100000000;
 static pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
-
-void Perror(const char *s)
-{
-    perror(s);
-    exit(EXIT_FAILURE);
-}
 
 long long getSystemTime()
 {
@@ -39,16 +36,20 @@ long long getSystemTime()
     return 1000 * t.time + t.millitm;
 }
 
-void* fun2(void *arg)
+void* fun(void *arg)
 {
-    pthread_t thread_id = pthread_self();
-    printf("The thread2 id is %ld\n", (long)thread_id);
     int i = 1;
     for (; i <= count; ++i) {
         pthread_mutex_lock(&mutex);
-        num++;
-        num *= 2;
+        num += 1;
+        printf("tid = %u, num = %d\n", pthread_self(), num);
         pthread_mutex_unlock(&mutex);
+
+        float s = 0;
+        int j;
+        for (j = 0; j < N; j++) {
+            s += sqrt(j);
+        }    
     }
 }
 
@@ -58,26 +59,24 @@ int main()
     pthread_t thread1;
     pthread_t thread2;
 
-    thread1 = pthread_self();
-    printf("The thread1 id is %ld\n", (long)thread1)
-
     long long start = getSystemTime();
 
     //create thread
-    err = pthread_create(&thread2, NULL, fun2, NULL);
+    err = pthread_create(&thread1, NULL, fun, NULL);
     if (err != 0) {
-        Perror("can't create thread2\n");
+        perror("can't create thread1\n");
+        exit(EXIT_FAILURE);
     }
 
-    int i = 1;
-    for (; i <= count; ++i) {
-        pthread_mutex_lock(&mutex);
-        num++;
-        num *= 2;
-        pthread_mutex_unlock(&mutex);
+    err = pthread_create(&thread2, NULL, fun, NULL);
+    if (err != 0) {
+        perror("can't create thread2\n");
+        exit(EXIT_FAILURE);
     }
-    
+
+    pthread_join(thread1, NULL);
     pthread_join(thread2, NULL);
+
     long long end = getSystemTime();
 
     printf("The num is %d, pay %lld ms\n", num, (end - start));
