--- Tuning_kit/testdemo/cache/cache_miss/miss_start.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-for i in {1..20};
-do
-	taskset -c 1 ./cache_miss;
-done
