--- Tuning_kit/testdemo/falsesharing/falsesharing_start.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-for i in {1..100}
-do
-	./falsesharing
-done
