--- Tuning_kit/testdemo/multiply/multiply_start.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-for i in {1..200}
-do
-	taskset -c 1 ./multiply
-done
